task 1

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -10,7 +10,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: float) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """ Asynchronous routine spawns wait_random n times with max_delay """
 
     tasks = [wait_random(max_delay) for _ in range(n)]
